Refactor location of repositories and table initiazliers
