+++ DAPM/DAPM.AccessControlService.Test.Unit/Repositories/ResourceRepositoryTests.cs
@@ -1,6 +1,6 @@
 using System.Data;
 using DAPM.AccessControlService.Core.Domain.Entities;
-using DAPM.AccessControlService.Infrastructure.Database;
+using DAPM.AccessControlService.Infrastructure.Repositories;
 using DAPM.AccessControlService.Test.Unit.Repositories.TableInitializers;
 using Microsoft.Data.Sqlite;
