+++ DAPM/DAPM.AccessControlService.Test.Unit/Services/ResourceServiceTests.cs
@@ -1,5 +1,5 @@
 using DAPM.AccessControlService.Core.Services;
-using DAPM.AccessControlService.Infrastructure.Database;
+using DAPM.AccessControlService.Infrastructure.Repositories;
 using DAPM.AccessControlService.Test.Unit.Repositories.TableInitializers;
 using Microsoft.Data.Sqlite;
 using RabbitMQLibrary.Models.AccessControl;
