--- DAPM/DAPM.AccessControlService/Infrastructure/Database/TableInitializers/ITableInitializer.cs
@@ -1,6 +1,0 @@
-namespace DAPM.AccessControlService.Infrastructure.Database.TableInitializers;
-
-public interface ITableInitializer<T>
-{
-    Task InitializeTable();
-}
