+++ DAPM/DAPM.AccessControlService/Infrastructure/TableInitializers/ITableInitializer.cs
@@ -1,0 +1,6 @@
+namespace DAPM.AccessControlService.Infrastructure.TableInitializers;
+
+public interface ITableInitializer<T>
+{
+    Task InitializeTable();
+}
